multirole access

+++ pages/pelanggan/pelanggan.php
@@ -58,9 +58,11 @@ include_once "../layout/header.php"
                             <!-- <h3 class="card-title">DataTable with default features</h3> -->
                             <!-- <a href="barang-tambah.php"><button type="button" class="btn btn-primary rounded">Tambah</button></a> -->
                             <!-- Button trigger modal -->
-                            <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#exampleModal">
-                                Tambah
-                            </button>
+                            <?php if ($_SESSION['role'] == 'karyawan') { ?>
+                                <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#exampleModal">
+                                    Tambah
+                                </button>
+                            <?php } ?>
                         </div>
                         <!-- /.card-header -->
                         <div class="card-body">
@@ -69,7 +71,9 @@ include_once "../layout/header.php"
                                     <tr>
                                         <th>ID Pelanggan</th>
                                         <th>Nama Pelanggan</th>
-                                        <th>Aksi</th>
+                                        <?php if ($_SESSION['role'] == 'karyawan') { ?>
+                                            <th>Aksi</th>
+                                        <?php } ?>
                                     </tr>
                                 </thead>
                                 <tbody>
@@ -80,16 +84,18 @@ include_once "../layout/header.php"
                                         <tr>
                                             <td><?= $item['id_pelanggan']; ?></td>
                                             <td><?= $item['nama_pelanggan']; ?></td>
-                                            <td>
-                                                <!-- a href -->
-                                                <a href="#" type="button" data-toggle="modal" data-target="#myModal<?= $item['id_pelanggan'] ?>" class="btn btn-success btn-circle btn-sm">
-                                                    <i class="fas fa-edit"></i>
-                                                </a>
-                                                <!-- a href -->
-                                                <a href="pelanggan-hapus.php?id_pelanggan=<?= $item['id_pelanggan']; ?>" class="btn btn-danger btn-circle btn-sm hapus">
-                                                    <i class="fas fa-trash"></i>
-                                                </a>
-                                            </td>
+                                            <?php if ($_SESSION['role'] == 'karyawan') { ?>
+                                                <td>
+                                                    <!-- a href -->
+                                                    <a href="#" type="button" data-toggle="modal" data-target="#myModal<?= $item['id_pelanggan'] ?>" class="btn btn-success btn-circle btn-sm">
+                                                        <i class="fas fa-edit"></i>
+                                                    </a>
+                                                    <!-- a href -->
+                                                    <a href="pelanggan-hapus.php?id_pelanggan=<?= $item['id_pelanggan']; ?>" class="btn btn-danger btn-circle btn-sm hapus">
+                                                        <i class="fas fa-trash"></i>
+                                                    </a>
+                                                </td>
+                                            <?php } ?>
                                         </tr>
                                         <!-- Modal Edit Data -->
                                         <div class="modal fade" id="myModal<?= $item['id_pelanggan'] ?>" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true" role="dialog">
